fix(合成视频): 1、增加进度、失败原因、完成时间获取

+++ yudao-module-digitalcourse/yudao-module-digitalcourse-biz/src/main/java/cn/iocoder/yudao/module/digitalcourse/controller/admin/coursemedia/vo/CourseMediaPageReqVO.java
@@ -41,7 +41,7 @@ public class CourseMediaPageReqVO extends PageParam {
     private String previewUrl;
 
     @Schema(description = "进度（预留）")
-    private Integer progress;
+    private Float progress;
 
     @Schema(description = "课程id", example = "9112")
     private Long courseId;
