+++ yudao-module-digitalcourse/yudao-module-digitalcourse-biz/src/main/java/cn/iocoder/yudao/module/digitalcourse/controller/admin/coursemedia/vo/CourseMediaSaveReqVO.java
@@ -31,7 +31,7 @@ public class CourseMediaSaveReqVO {
     private String previewUrl;
 
     @Schema(description = "进度（预留）")
-    private Integer progress;
+    private Float progress;
 
     @Schema(description = "课程id", example = "9112")
     private Long courseId;
